K23CNT1_NguyenAnhTuan_2310900113_Project1

+++ K23CNT1_NguyenAnhTuan_2310900113_Project1/database/migrations/2024_12_18_142631_create__n_a_t__c_t__h_o_a__d_o_n_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->bigInteger('natSanPhamID')->references('id')->on('NAT_SAN_PHAM');
             $table->integer('natSoLuongMua');  // Số lượng mua
             $table->float('natDonGiaMua');  // Đơn giá mua
-            $table->float('natThanhTien');  // Thành tiền
+            $table->double('natThanhTien');  // Thành tiền
             $table->tinyInteger('natTrangThai'); 
             $table->timestamps();
         });
